Changes:
* better instances
  - implement `Ord [a]` by defining `compare` instead of `<=`
  - define `*>`, `liftA2` in the `Applicative []` instance
  - define `stimes` in the `Semigroup [a]` instance
* optimizations
  - ~~define `length` the naive way, in my micro-benchmark (`print $ length [1..10_000::Int]`) this was a lot better (taking 2/3 of the reductions)~~
  - optimize `span`, `spanUntil`, `splitWith`, `inits` to avoid `reverse`
  - optimize `partition`, `unsnoc` to use a single traversal
  - use merge sort for sorting (see #74)
  - inline `combine` in `transpose`
  - add shortcuts for empty lists in `intersectBy`
  - slightly better `unlines`, `unwords` (using the `base` versions)
* avoid helpers for functions that are optimized anyway (`concatMap`, `map`, `filter`, `foldr`, `foldr1`, `foldr1'`, `scanl`, `scanl'`, `scanr`, `scanr1`), I checked that the resulting code is the same by comparing the combinator files
* improve error messages (taken from `base`)
* remove unexported `*By` functions (this made basically no performance difference)
* use `break ...` instead of `span (not . ...)`
* name recursive helpers `go` instead of `rec`, since `rec` is a keyword in GHC (with `RecursiveDo`) and `go` is a common name for recursive helpers