@konsumlamm 2025-01-24
Turns out the naive `length` definition leads to a stack overflow when building the compiler.