@augustss 2025-01-24
I don't like the name 'go' even if that is very common.  :)

Did you check that the new versions of span and unsnoc are actually better?