@konsumlamm 2025-01-24
> I don't like the name 'go' even if that is very common. :)

What name would you prefer? `rec` is highlighted as a keyword in my editor, which is a bit annoying.

> Did you check that the new versions of span and unsnoc are actually better?

I did a few micro-benchmarks which showed less allocations and reductions, although it's possible that they're not representative of real use cases.

<details>

#### unsnoc
```hs
import Control.DeepSeq (force)
import Control.Exception (evaluate)
import Data.List

main :: IO ()
main = do
    _ <- evaluate $ force $ unsnoc [1..10_000::Int]
    pure ()
```
old:
```
         10.410 combinator file size
          1.989 cells at start
     50.000.000 cells heap size (800.000.000 bytes)
        363.666 cells allocated (2,8 Mbyte/s)
              2 GCs
          1.989 max cells used
      1.383.744 reductions (0,7 Mred/s)
              1 array alloc
              0 array free
              0 foreign alloc
              0 foreign free
             23 bytestring alloc (max 22)
            228 bytestring alloc bytes (max 228)
             23 bytestring free
           2,10s total expired time
           0,00s total gc time (0,00 + 0,00)
 GC reductions A=243, K=0, I=0, int=0 flip=0
 special reductions B'=0 K4=0 K3=0 K2=0 C'B=0, Z=1, R=0
```
new:
```
         10.274 combinator file size
          1.969 cells at start
     50.000.000 cells heap size (800.000.000 bytes)
        343.641 cells allocated (2,7 Mbyte/s)
              2 GCs
          1.969 max cells used
      1.203.744 reductions (0,6 Mred/s)
              1 array alloc
              0 array free
              0 foreign alloc
              0 foreign free
             21 bytestring alloc (max 20)
            196 bytestring alloc bytes (max 196)
             21 bytestring free
           2,05s total expired time
           0,00s total gc time (0,00 + 0,00)
 GC reductions A=241, K=0, I=0, int=0 flip=0
 special reductions B'=0 K4=0 K3=0 K2=0 C'B=0, Z=1, R=0
```

#### span
```hs
import Control.DeepSeq (force)
import Control.Exception (evaluate)
import Data.List

main :: IO ()
main = do
    _ <- evaluate $ force $ span (< 500) [1..10_000::Int]
    pure ()
```
old:
```
         10.338 combinator file size
          1.987 cells at start
     50.000.000 cells heap size (800.000.000 bytes)
        209.241 cells allocated (1,4 Mbyte/s)
              2 GCs
          1.987 max cells used
        782.017 reductions (0,3 Mred/s)
              1 array alloc
              0 array free
              0 foreign alloc
              0 foreign free
             21 bytestring alloc (max 20)
            196 bytestring alloc bytes (max 196)
             21 bytestring free
           2,31s total expired time
           0,00s total gc time (0,00 + 0,00)
 GC reductions A=240, K=0, I=0, int=0 flip=0
 special reductions B'=0 K4=0 K3=0 K2=0 C'B=0, Z=0, R=0
```
new:
```
         10.288 combinator file size
          1.978 cells at start
     50.000.000 cells heap size (800.000.000 bytes)
        205.721 cells allocated (1,5 Mbyte/s)
              2 GCs
          1.978 max cells used
        781.491 reductions (0,4 Mred/s)
              1 array alloc
              0 array free
              0 foreign alloc
              0 foreign free
             21 bytestring alloc (max 20)
            196 bytestring alloc bytes (max 196)
             21 bytestring free
           2,20s total expired time
           0,00s total gc time (0,00 + 0,00)
 GC reductions A=239, K=0, I=0, int=0 flip=0
 special reductions B'=0 K4=0 K3=0 K2=0 C'B=0, Z=0, R=0
```

</details>

> Completely unrelated: With the name @konsumlamm I'm guessing that you're Swedish? :)

I'm actually german :)