@augustss 2025-01-24
Thanks for the performance details.

I don't have a good alternative name for 'go', so lets use that.
I prefer 'rec', but as you say, it is reserved with the RecursiveDo extension.